requirements.txt
telethon
python-dotenv
aiohttp
cryptg
pillow